Introduce a side menu for each issue
